backen-michelle/src/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
sqlalchemy==2.0.27
aiomysql==0.2.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
openai==1.12.0
SpeechRecognition==3.10.1
alembic==1.13.1
bcrypt==4.1.2
pytest==8.0.1
pytest-asyncio==0.23.5
httpx==0.26.0